Add function to compute gap penalties used in the WT-ALT peptides alignment scoring
A gap function was added to favor alignments that have their gaps appear later in the
sequence (more to the right). This is useful to avoid cases where multiple alignments
have the same score, yet are only distinguished by the starting position of a sequence
of gaps, e.g.:

1-
CGGHGGPRAASSSPSCPGCGPPGPGAHCPGVPRSAPAQAPTSGADLSAHLW-----
|||||||||||||||||||||||     |||||||||||||||||||||||
CGGHGGPRAASSSPSCPGCGPPG-----PGVPRSAPAQAPTSGADLSAHLWARYQD
  Score=45.1

2-
CGGHGGPRAASSSPSCPGCGPPGPGAHCPGVPRSAPAQAPTSGADLSAHLW-----
||||||||||||||||||||||||     ||||||||||||||||||||||
CGGHGGPRAASSSPSCPGCGPPGP-----GVPRSAPAQAPTSGADLSAHLWARYQD
  Score=45.1

3-
CGGHGGPRAASSSPSCPGCGPPGPGAHCPGVPRSAPAQAPTSGADLSAHLW-----
|||||||||||||||||||||||||     |||||||||||||||||||||
CGGHGGPRAASSSPSCPGCGPPGPG-----VPRSAPAQAPTSGADLSAHLWARYQD
  Score=45.1

We want the keep the last alignment because this allows us to (later in the code) report
the correct positions (in terms of order) in the ALT sequence where the mutations occur.
E.g, in the example above, keeping alignment 1 would report positions 24-27, whereas
alignment 3 reports the correct positions for the mutations in the ALT peptide: 26-30.

Also,
Add parameter to global alignment function that only penalizes gaps when they are found in the ALT peptide sequence.

There was no case where this was seen to make a difference, but conceptually it seems to
make sense for our use case. Looking at the example above, we don't want to penalize
alignments where the ALT sequence ends up getting "longer" than the WT peptide just because
gaps were inserted in the middle of the sequence. On the other hand we want to penalize
the inverse case, as we'd rather have mismatches in the end then gaps for the ALT peptide.