Fix incorrect mutation position handling case in ALT peptide
The case happens when the ALT alignment sequence starts with
gaps. An offset was addedto counter this.
This bug also affected the peptides generated for BLAST, i.e.,
peptides that did not have the mutation were being generated.
Example where it was incorrect:
Output in file: chr22_50508443_G/A_M26,44

Alignment being done:
	                         *                 *
	LLQPIRDLTKNWEVDVAAQLGEYLEELDQICISFDEGKTTMNFIEAALLIQ
	---PIRDLTKNWEVDVAAQLGEYLEKLDQICISFDEGKTTMNFLEAALLIQ
						 EYLEKLDQ          TMNFLEAA
						  YLEKLDQI          MNFLEAAL
                           LEKLDQIC          NFLEAALL
                            EKLDQICI          FLEAALLI
                             KLDQICIS          LEAALLIQ
                              LDQICISF
                               DQICISFD
                                QICISFDE

Output in file should have been: chr22_50508443_G/A_M23,41
Also note that the last 3 short peptides from the mutation in pos 23
did not include the mutation. The reason why the mutation in
pos 44 has not generated "wrong" peptides is because the sliding
window reaches the end of the sequence, as such there are no
more small peptides to generate.