Filter out unnecessary peptides from being blasted/binding predicted
This was done taking into account the observation that the helper
function "take_sub_peptide" followed by "calc_mers" (used to compute
the short peptides that are then blasted and then go through binding
prediction) output peptides that do not include the mutation position
in the input peptide sequence. This happened, e.g., in the case where
the mutation was close to the end of the sequence, see:

chrX_138632861_C/T                              *
LPKPLKVAMYKEPSLHDLTEFSRSGSGTPTKSRSVSGVLNGGKSMSHNEST
LPKPLKVAMYKEPSLHDLTEFSRSGSGTPTKSRSVSGVLNGGKSMSHNKST
                                    GVLNGGKS
                                     VLNGGKSM
                                      LNGGKSMS
                                       NGGKSMSH
                                        GGKSMSHN

Above we see short peptides of size 8 that were computed and then blasted
and went through binding prediction unnecessarily since they do not
include the mutation (marked with *).