Add logic to handle empty WT variants (e.g. stop loss)
To do this, in the first part of the script - when computing the short
peptides (that are then blasted and enter binding prediction) -  we skip
pairwise alignment and instead inject a false WT peptide comprised of
"X"s (chosen because it is the any/unknown character for amino acids)
to the corresponding 'WT_pep' variable and adapt the remaining ones:
'WT_offset', 'MUT_offset' and 'MUT_WT_pos', allowing us to keep the
script's current logic. The other adjustment required is checking
in the binding prediction if we're upon a case of an empty WT, so that
we don't do prediction to the "XXXXX..." sequences.